requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
black==20.8b1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
ecdsa==0.14.1
fastapi==0.63.0
flake8==3.8.3
h11==0.9.0
httptools==0.1.1
idna==2.10
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.1
pandas==1.2.3
passlib==1.7.4
pathspec==0.8.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.6.1
pyflakes==2.2.0
python-dateutil==2.8.1
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
regex==2020.7.14
requests==2.24.0
rsa==4.6
six==1.15.0
starlette==0.13.6
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
